Processed review comments.
